backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pytesseract>=0.3.10
Pillow>=10.2.0
pdf2image>=1.17.0
opencv-python-headless>=4.10.0.84
pydantic>=2.7.0
